+++ pdata/internal/cmd/pdatagen/internal/base_fields.go
@@ -62,12 +62,13 @@ const copyToPrimitiveSliceTestTemplate = `	if len(ms.getOrig().${originFieldName
 
 const accessorsPrimitiveSliceTemplate = `// ${fieldName} returns the ${lowerFieldName} associated with this ${structName}.
 func (ms ${structName}) ${fieldName}() ${packageName}${returnType} {
-	return ${packageName}${returnType}(internal.New${returnType}(ms.getOrig().${originFieldName}))
+	return ${packageName}${returnType}(internal.New${returnType}(&ms.getOrig().${originFieldName}))
 }
 
 // Set${fieldName} replaces the ${lowerFieldName} associated with this ${structName}.
+// Deprecated: [0.60.0] Use ${fieldName}().FromRaw() instead
 func (ms ${structName}) Set${fieldName}(v ${packageName}${returnType}) {
-	ms.getOrig().${originFieldName} = internal.GetOrig${returnType}(internal.${returnType}(v))
+	ms.getOrig().${originFieldName} = *internal.GetOrig${returnType}(internal.${returnType}(v))
 }`
 
 const oneOfTypeAccessorHeaderTemplate = `// ${originFieldName}Type returns the type of the ${lowerOriginFieldName} for this ${structName}.
@@ -171,10 +172,9 @@ const accessorsPrimitiveTypedTestTemplate = `func Test${structName}_${fieldName}
 
 const accessorsPrimitiveSliceTestTemplate = `func Test${structName}_${fieldName}(t *testing.T) {
 	ms := New${structName}()
-	assert.Equal(t, ${packageName}New${returnType}(${defaultVal}), ms.${fieldName}())
-	testVal${fieldName} := ${packageName}New${returnType}(${testValue})
-	ms.Set${fieldName}(testVal${fieldName})
-	assert.Equal(t, testVal${fieldName}, ms.${fieldName}())
+	assert.Equal(t, ${defaultVal}, ms.${fieldName}().AsRaw())
+	ms.${fieldName}().FromRaw(${testValue})
+	assert.Equal(t, ${testValue}, ms.${fieldName}().AsRaw())
 }`
 
 const accessorsOptionalPrimitiveValueTemplate = `// ${fieldName} returns the ${lowerFieldName} associated with this ${structName}.